+++ CHANGES.md
@@ -1,6 +1,9 @@
 ## unreleased
 
-- Improve documentation.
+- Changed the processing instruction API to not use `get_target` and
+  `get_value` but instead to use `target()` and `value()`.
+
+- Documentation tweaks.
 
 - Tweak `Cargo.toml`